module/home/firefox: check if firefox is running

+++ module/home/firefox/default.nix
@@ -373,6 +373,17 @@ in
 
             cat <<\AQUARIS_FIREFOX_EOF >> tmp
             export FIREFOX_PROFILE_DIR="$HOME/${dir}/"
+
+            # export whether firefox is already running for pre/postRun scripts
+            # 1. if the permissions DB does not exist: before first run -> not running
+            # 2. otherwise, if we can open it: not locked -> not running
+            # 3. otherwise: running
+            if [ ! -e "$FIREFOX_PROFILE_DIR/permissions.sqlite" ] ||
+              ${getExe pkgs.sqlite} -readonly           \
+              "$FIREFOX_PROFILE_DIR/permissions.sqlite" \
+              ".schema" >/dev/null; then RUNNING=0; else RUNNING=1; fi
+            export RUNNING
+
             ${cfg.preRun}
             AQUARIS_FIREFOX_EOF
 
@@ -439,11 +450,13 @@ in
         preRun = pipe cfg.sanitize.exceptions [
           (map (url: "('${url}', 'cookie', 1, 0, 0)"))
           (x: ''
-            ${getExe pkgs.sqlite} "$FIREFOX_PROFILE_DIR/permissions.sqlite" << EOF
-            delete from moz_perms where type = 'cookie';
-            insert into moz_perms (origin, type, permission, expireType, expireTime) values
-            ${builtins.concatStringsSep ",\n" x};
-            EOF
+            if ((RUNNING == 0)); then
+            	${getExe pkgs.sqlite} "$FIREFOX_PROFILE_DIR/permissions.sqlite" <<-EOF
+            		delete from moz_perms where type = 'cookie';
+            		insert into moz_perms (origin, type, permission, expireType, expireTime) values
+            		${builtins.concatStringsSep ",\n\t\t" x};
+            	EOF
+            fi
           '')
         ];
       };
